refactor to new file, add test initialization
